totally redid the javascript
Broke it down into sections. First I had to make sure the search button was working and that data was getting pulled in from the API. Then had to make sure the right data was getting pulled. Once that happened, I realized that to get the windspeed and UV index for the forecast, I would have to use a different API call. Took a lot of trial and error to get that set up and working. Since they were firing off at the same time, I had to change use setTimeout to create a  slight delay to avoid the clash. After data was being pulled from the API, I than had to ensure that it was being connected to and displayed in the appropriate elements of the DOM. Once that was finally working, I had to then ensure it was logging to local storage. Used the inspector to check it along the way  to ensure it was logging.